# docs/swagger.yaml
basePath: /
definitions:
  response.ErrResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
  v1.signInRequest:
    properties:
      email:
        example: email@example.com
        maxLength: 150
        minLength: 5
        type: string
      password:
        example: YourV@lidPassw0rd!
        maxLength: 32
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  v1.signInResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  v1.signUpRequest:
    properties:
      email:
        example: email@example.com
        maxLength: 150
        minLength: 5
        type: string
      password:
        example: YourV@lidPassw0rd!
        maxLength: 32
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  v1.signUpResponse:
    properties:
      id:
        example: d13a75e2-3d21-4e57-9dc0-3a7f5bee4c25
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Authorization, registration, etc...
  title: Universal authorization service API
  version: "1.0"
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Sign in
      parameters:
      - description: Sign in payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.signInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.signInResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Sign in
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Sign up
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.signUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.signUpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Sign up
      tags:
      - auth
swagger: "2.0"